editor style fixes

+++ client/src/components/NotebookList.jsx
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from 'react';
 import NotebookItem from './NotebookItem';
 import { getCategories, createNotebook, deleteNotebook } from '../api'
 
-function NotebookList({onSelectNotebook}) {
+function NotebookList({onSelectNotebook, selectedNotebook}) {
 
  const [notebooks, setNotebooks] = useState([]);
  const[newNotebookName, setNewNotebookName] = useState('')
@@ -31,6 +31,7 @@ const handleCreateNotebook = async(event) => {
 }
 
 const handleDeleteNotebook = async (notebookId) => {
+  if (!window.confirm('Эта категория будет удалена безвозвратно со всеми принадлежащими ей заметками. Действительно хотите ее удалить?')) return;
   try {
     await deleteNotebook(notebookId);
     setNotebooks(notebooks.filter(notebook => notebook.id !== notebookId))
@@ -43,22 +44,24 @@ const handleDeleteNotebook = async (notebookId) => {
 
 
   return (
-    <div>
+    <div className='notebook_list'>
       <h2>Категории</h2>
-      <button onClick={() => setShowInput(true)}>Добавить</button>
+      <p className='editor_hint' style={{width: '93%', fontSize: '12px'}}>Выберите или создайте категорию для заметок</p>
+      <button onClick={() => setShowInput(true)} style={{width: '93%'}}>Добавить</button>
       {showInput && (
         <form onSubmit={handleCreateNotebook}>
           <input 
             type="text"
             value={newNotebookName}
             onChange={(event) => setNewNotebookName(event.target.value)}
-            placeholder="Название новой рубрики"
+            placeholder="Название категории"
+            style={{width: '93%'}}
           />
-          <button type="submit">Создать</button>
+          <button type="submit" style={{width: '93%'}}>Создать</button>
         </form>
       )}
       {notebooks.map((notebook) => (
-        <NotebookItem key={notebook.id} notebook={notebook} onSelect={onSelectNotebook} onDelete={handleDeleteNotebook}/>
+        <NotebookItem key={notebook.id} notebook={notebook} onSelect={onSelectNotebook} onDelete={handleDeleteNotebook} isSelected={selectedNotebook && notebook.id === selectedNotebook.id}/>
       ))}
     </div>
   );
